# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "omit to start a new conversation",
            "title": "Conversation Id"
          },
          "focus_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Graph node key to pull a blast radius into the context. Set by the graph panel when the user asks about a selected node; there is no intent detection on the message itself",
            "title": "Focus Key"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "GraphNode": {
        "properties": {
          "change_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Commits touching this file, once history has been read",
            "title": "Change Count"
          },
          "display": {
            "description": "What to label the node with",
            "title": "Display",
            "type": "string"
          },
          "is_external": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is External"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "function | method, on :Function",
            "title": "Kind"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line End"
          },
          "line_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Start"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "qualname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualname"
          },
          "type": {
            "description": "Repo | File | Class | Function | Module",
            "title": "Type",
            "type": "string"
          },
          "unresolved_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unresolved Calls"
          }
        },
        "required": [
          "key",
          "type",
          "display"
        ],
        "title": "GraphNode",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImpactExplanationResponse": {
        "description": "The blast radius in prose.\n\nThe counts travel with the text so a reader can see what it was derived from\n\u2014 an explanation of the top 25 of 300 affected nodes is a different claim\nfrom an explanation of all 12.",
        "properties": {
          "affected": {
            "description": "Nodes in the radius",
            "title": "Affected",
            "type": "integer"
          },
          "cached": {
            "description": "True when served from the per-process cache",
            "title": "Cached",
            "type": "boolean"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "explained": {
            "description": "Of those, how many were described to the model",
            "title": "Explained",
            "type": "integer"
          },
          "input_tokens": {
            "default": 0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "model": {
            "description": "Empty when no model was called",
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "default": 0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "refused": {
            "default": false,
            "description": "True when the provider declined the request",
            "title": "Refused",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "truncated": {
            "description": "True when the cap hid part of the radius",
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "display",
          "text",
          "model",
          "affected",
          "explained",
          "depth",
          "truncated",
          "cached"
        ],
        "title": "ImpactExplanationResponse",
        "type": "object"
      },
      "ImpactResponse": {
        "properties": {
          "items": {
            "description": "Ranked by score, highest first",
            "items": {
              "$ref": "#/components/schemas/ImpactedNode"
            },
            "title": "Items",
            "type": "array"
          },
          "root": {
            "$ref": "#/components/schemas/GraphNode"
          },
          "summary": {
            "$ref": "#/components/schemas/ImpactSummary"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "truncated": {
            "description": "True when the cap hid part of the radius",
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "root",
          "items",
          "total",
          "truncated",
          "summary"
        ],
        "title": "ImpactResponse",
        "type": "object"
      },
      "ImpactSummary": {
        "properties": {
          "by_hop": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Affected node count per hop distance",
            "title": "By Hop",
            "type": "object"
          },
          "by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Affected node count per node type",
            "title": "By Type",
            "type": "object"
          },
          "decay": {
            "title": "Decay",
            "type": "number"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "direct": {
            "description": "Nodes one hop out \u2014 the ones most likely to break",
            "title": "Direct",
            "type": "integer"
          },
          "max_hops": {
            "title": "Max Hops",
            "type": "integer"
          },
          "top_score": {
            "title": "Top Score",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "depth",
          "decay",
          "direct",
          "max_hops",
          "top_score",
          "by_hop",
          "by_type"
        ],
        "title": "ImpactSummary",
        "type": "object"
      },
      "ImpactedNode": {
        "description": "A node a change to the target could reach.",
        "properties": {
          "change_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Commits touching this file, once history has been read",
            "title": "Change Count"
          },
          "confidence": {
            "description": "Product of per-hop edge confidences",
            "title": "Confidence",
            "type": "number"
          },
          "display": {
            "description": "What to label the node with",
            "title": "Display",
            "type": "string"
          },
          "hops": {
            "description": "Edges between the target and this node",
            "title": "Hops",
            "type": "integer"
          },
          "is_external": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is External"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "function | method, on :Function",
            "title": "Kind"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line End"
          },
          "line_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Start"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "qualname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualname"
          },
          "route": {
            "description": "Node keys from the target to this node, inclusive of both. Named `route` because `path` is already this node's file path.",
            "items": {
              "type": "string"
            },
            "title": "Route",
            "type": "array"
          },
          "score": {
            "description": "confidence x decay^(hops-1) \u2014 the ranking key",
            "title": "Score",
            "type": "number"
          },
          "type": {
            "description": "Repo | File | Class | Function | Module",
            "title": "Type",
            "type": "string"
          },
          "unresolved_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unresolved Calls"
          },
          "via": {
            "description": "Edge types along the route, target first",
            "items": {
              "type": "string"
            },
            "title": "Via",
            "type": "array"
          }
        },
        "required": [
          "key",
          "type",
          "display",
          "hops",
          "confidence",
          "score",
          "via",
          "route"
        ],
        "title": "ImpactedNode",
        "type": "object"
      },
...
      "ParseStatus": {
        "enum": [
          "pending",
          "parsing",
          "complete",
          "failed"
        ],
        "title": "ParseStatus",
        "type": "string"
      },
...
      "RepositoryOut": {
        "properties": {
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Branch"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "file_count": {
            "title": "File Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parsed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed At"
          },
          "status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "symbol_count": {
            "title": "Symbol Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "default_branch",
          "commit_sha",
          "status",
          "error_message",
          "file_count",
          "symbol_count",
          "created_at",
          "updated_at",
          "parsed_at"
        ],
        "title": "RepositoryOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI Engineering Intelligence Platform",
    "title": "ARGUS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/repos/{repository_id}": {
      "delete": {
        "operationId": "delete_repository_repos__repository_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Repository",
        "tags": [
          "repositories"
        ]
      },
      "get": {
        "operationId": "get_repository_repos__repository_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/repos/{repository_id}/chat": {
      "post": {
        "description": "Ask a question. Streams SSE: `context`, then `delta`s, then `done`.",
        "operationId": "chat_repos__repository_id__chat_post",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/repos/{repository_id}/impact": {
      "get": {
        "description": "The ranked blast radius of a change to `key`, with the path to each node.",
        "operationId": "get_impact_repos__repository_id__impact_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "description": "Node key, from /graph/search",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "description": "Node key, from /graph/search",
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 5,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Score retained per hop; 1 disables decay",
            "in": "query",
            "name": "decay",
            "required": false,
            "schema": {
              "default": 0.6,
              "description": "Score retained per hop; 1 disables decay",
              "exclusiveMinimum": 0,
              "maximum": 1,
              "title": "Decay",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Impact",
        "tags": [
          "impact"
        ]
      }
    },
    "/repos/{repository_id}/impact/explain": {
      "get": {
        "description": "The same blast radius as `/impact`, explained in prose.\n\nNot streamed, unlike `/chat`: this is a few hundred tokens rendered beside a\ngraph, and a single JSON response is cacheable, re-renderable and far easier\nfor the panel to hold than a stream it would have to reassemble.",
        "operationId": "explain_repos__repository_id__impact_explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "description": "Node key, from /graph/search",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "description": "Node key, from /graph/search",
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 5,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactExplanationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain",
        "tags": [
          "impact"
        ]
      }
    },
...
  }
}